Use Kazv::Store in place of lager::store
Kazv::Store uses a BoolPromiseInterface and returns a
BoolPromise in dispatch(). The BoolPromise will be
resolved when all of the following are met:

(1) The reducer for that action has been called;
(2) The effect for that action has been called, if any;
(3) If the effect returns a BoolPromise, that promise
    is resolved.

For example, if you dispatch an Action, whose reducer is
like:

    Result update(Model m, Action a) {
        return {
            someProcessing(m),
            [](auto &&ctx) {
                return ctx.dispatch(AnotherAction{});
            }
        };
    }

Then the resulting promise of `auto p1 = ctx.dispatch(Action{});`
will be resolved *at least* after the reducer of AnotherAction
has been run. If the effect of AnotherAction dispatches any other
action, `p1` will be resolved after the reducer of those actions
has been run, recursively.