Return the Promise by effect if there are no jobs
We used to return `ph.all(promises by jobs)` in SdkModel::update
but that is a bug as it did not consider the actions that does
not have a job (namely, ProcessResponseAction). In that case, the
Promise returned by the reducer for that action was simply discarded.
This causes problem for Client::uploadContent, which returns useful data
via the Promise.