Optimise sortedUniqueMerge()
Now the sortedUniqueMerge() does a binary search for every addon item
in the original list and insert it there.

If there are n incoming events, and originally N events in the timeline,
the complexity is effectively O(n log(N)). Since n <= (incremental sync
limit), it is effectively O(log(N)).